requirements.txt
aistudio-sdk==0.3.8
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
bce-python-sdk==0.9.55
cachetools==6.2.2
certifi==2025.11.12
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
colorlog==6.10.1
cssselect==1.3.0
cssutils==2.11.1
distro==1.9.0
dnspython==2.8.0
einops==0.8.1
email-validator==2.3.0
et_xmlfile==2.0.0
exceptiongroup==1.3.1
fastapi==0.124.0
fastapi-cli==0.0.16
fastapi-cloud-cli==0.6.0
fastar==0.8.0
filelock==3.20.0
fsspec==2025.12.0
ftfy==6.3.1
future==1.0.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.2.1
idna==3.11
imagesize==1.4.1
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.2
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
modelscope==1.32.0
more-itertools==10.8.0
networkx==3.4.2
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-nvjitlink-cu12==12.9.86
openai==2.9.0
opencv-contrib-python==4.10.0.84
openpyxl==3.1.5
opt-einsum==3.3.0
packaging==25.0
paddleocr==3.3.2
paddlepaddle-gpu==3.2.1
paddlex==3.3.10
pandas==2.3.3
pdf2image==1.17.0
pillow==11.2.1
premailer==3.10.0
prettytable==3.17.0
protobuf==6.31.1
psutil==7.1.3
py-cpuinfo==9.0.0
pyclipper==1.4.0
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyMuPDF==1.26.6
pypdfium2==5.1.0
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
regex==2025.11.3
requests==2.32.5
rich==14.2.0
rich-toolkit==0.17.0
rignore==0.7.6
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.15
safetensors @ https://xly-devops.cdn.bcebos.com/safetensors-nightly/safetensors-0.6.2.dev0-cp38-abi3-win_amd64.whl#sha256=e829574729a18cb4609ad11c8df9c1b810ca1b98540edae48f4d4ed6f103e012
scikit-learn==1.7.2
scipy==1.15.3
sentencepiece==0.2.1
sentry-sdk==2.47.0
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.50.0
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
tomli==2.3.0
tqdm==4.67.1
typer==0.20.0
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.11.0
urllib3==2.6.0
uvicorn==0.38.0
watchfiles==1.1.1
wcwidth==0.2.14
websockets==15.0.1